Consolidate sceneItem component
