--- app-frontend/src/app/components/scenes/sceneItem/sceneItem.component.js
@@ -1,23 +1,0 @@
-// Component code
-import sceneTpl from './sceneItem.html';
-
-const rfSceneItem = {
-    templateUrl: sceneTpl,
-    transclude: true,
-    controller: 'SceneItemController',
-    bindings: {
-        scene: '<',
-        actions: '<',
-        onView: '&',
-        onDownload: '&',
-
-        selected: '<?',
-        onSelect: '&?',
-        isDisabled: '<?',
-        apiSource: '<?',
-        planetKey: '<?',
-        onPassPlanetThumbnail: '&?'
-    }
-};
-
-export default rfSceneItem;
