+++ app-frontend/src/app/pages/projects/edit/browse/browse.html
@@ -62,10 +62,7 @@
       scene="scene"
       selected="$ctrl.isSelected(scene)"
       is-disabled="$ctrl.isInProject(scene)"
-      api-source="$ctrl.sourceRepo"
-      planet-key="$ctrl.planetKey"
       on-select="$ctrl.setSelected(scene, selected)"
-      on-pass-planet-thumbnail="$ctrl.onPassPlanetThumbnail(url, id)"
       ng-click="$ctrl.openSceneDetailModal(scene)"
       ng-mouseenter="$ctrl.setHoveredScene(scene)"
       ng-mouseleave="$ctrl.removeHoveredScene()"
